Wrap Preview state helper functions into PreviewStateHelper class
