+++ WordPress/src/main/java/org/wordpress/android/ui/ActivityLauncherWrapper.kt
@@ -2,11 +2,11 @@ package org.wordpress.android.ui
 
 import android.app.Activity
 import android.content.Context
+import dagger.Reusable
 import org.wordpress.android.fluxc.model.PostModel
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.ui.posts.RemotePreviewLogicHelper.RemotePreviewType
 import javax.inject.Inject
-import javax.inject.Singleton
 
 /**
  * Injectable wrapper around ActivityLauncher.
@@ -15,7 +15,7 @@ import javax.inject.Singleton
  * Main purpose of this wrapper is to make testing easier.
  *
  */
-@Singleton
+@Reusable
 class ActivityLauncherWrapper @Inject constructor() {
     fun showActionableEmptyView(
         context: Context,
