+++ WordPress/src/main/java/org/wordpress/android/viewmodel/pages/PagesViewModel.kt
@@ -41,8 +41,8 @@ import org.wordpress.android.ui.pages.PageItem.Page
 import org.wordpress.android.ui.pages.SnackbarMessageHolder
 import org.wordpress.android.ui.posts.PostInfoType
 import org.wordpress.android.ui.posts.PostListRemotePreviewState
+import org.wordpress.android.ui.posts.PreviewStateHelper
 import org.wordpress.android.ui.posts.RemotePreviewLogicHelper.RemotePreviewType
-import org.wordpress.android.ui.posts.managePreviewStateTransitions
 import org.wordpress.android.ui.uploads.PostEvents
 import org.wordpress.android.ui.uploads.UploadStarter
 import org.wordpress.android.util.AppLog
@@ -89,6 +89,7 @@ class PagesViewModel
     private val networkUtils: NetworkUtilsWrapper,
     private val uploadStarter: UploadStarter,
     private val eventBusWrapper: EventBusWrapper,
+    private val previewStateHelper: PreviewStateHelper,
     @Named(UI_THREAD) private val uiDispatcher: CoroutineDispatcher,
     @Named(BG_THREAD) private val defaultDispatcher: CoroutineDispatcher
 ) : ScopedViewModel(uiDispatcher) {
@@ -313,7 +314,12 @@ class PagesViewModel
         _previewState.postValue(newState)
 
         // take care of exit actions on state transition
-        managePreviewStateTransitions(newState, prevState, postInfo, this::handleRemotePreview)
+        previewStateHelper.managePreviewStateTransitions(
+                newState,
+                prevState,
+                postInfo,
+                this::handleRemotePreview
+        )
     }
 
     private suspend fun groupedSearch(
